+++ .travis/deps.sh
@@ -3,5 +3,5 @@
 if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
     eval "$(curl -sL https://swiftenv.fuller.li/install.sh)";
 else
-    gem install cocoapods --pre --no-rdoc --no-ri --no-document --quiet;
+    gem install cocoapods --no-document --quiet;
 fi