+++ Sources/Codable/KeyedDecodingContainer+FailableSequence.swift
@@ -12,9 +12,22 @@ extension KeyedDecodingContainer {
     /// - throws: `DecodingError.keyNotFound` if `self` does not have an entry for the given key.
     /// - throws: `DecodingError.valueNotFound` if `self` has a null entry for the given key.
     public func decode<T>(forKey key: K, allowInvalidElements: Bool = false) throws -> [T] where T: Decodable {
-        guard allowInvalidElements else { return try decode([T].self, forKey: key) } // Standard decoding on none failable elements
-        let optionals = try decode([T?].self, forKey: key) // Decodes each element optionally
-        return optionals.compactMap { $0 } // Flatten-out any failed object
+        // Standard decoding on none failable elements
+        guard allowInvalidElements else {
+            return try decode([T].self, forKey: key)
+        }
+
+        // Directly read from unkeyed container behind given `key`
+        var container = try nestedUnkeyedContainer(forKey: key)
+        guard let numberOfElements = container.count else {
+            throw DecodingError.dataCorruptedError(in: container, debugDescription: "Container of unknown lenght")
+        }
+
+        // Decodes each element optionally
+        return stride(from: 0, to: numberOfElements, by: 1).compactMap({ _ in
+            guard !container.isAtEnd else { return nil }
+            return try? container.decode(T.self)
+        })
     }
 
     /// Decodes a optional sequence of a given type.
@@ -27,8 +40,7 @@ extension KeyedDecodingContainer {
     /// - throws: `DecodingError.keyNotFound` if `self` does not have an entry for the given key.
     /// - throws: `DecodingError.valueNotFound` if `self` has a null entry for the given key.
     public func decode<T>(forKey key: K, allowInvalidElements: Bool = false) throws -> [T]? where T: Decodable {
-        guard allowInvalidElements else { return try decodeIfPresent([T].self, forKey: key) } // Standard decoding on none failable elements
-        let optionals = try decodeIfPresent([T?].self, forKey: key)
-        return optionals?.compactMap { $0 } // Flatten-out any failed object
+        guard contains(key) else { return nil }
+        return try decode(forKey: key, allowInvalidElements: allowInvalidElements) as [T]
     }
 }