+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# 5.0
+# 5.1
+
+- Bumps swift version (swift 5.0)
+- Remove `Result` in favor of Swift built-in `Result` Type
+- Add storyboard identifier extension to `UIViewController`
+
+## 5.0
+
 - Bumps swift version (swift 4.2)
-# 4.1.0
+
+## 4.1.0
+
 - Initial Public Release